rename file with champion list related actions
removes ambiguity in action references
